refactor(azure-quickstart): move status reporter from azure quickstart to shared
